+++ Pengo2/Environment.cpp
@@ -440,7 +440,7 @@ void Environment::EnemyPlayerBreakBlock()
 {
 	GameEngine::HitInfo hitInfo;
 
-	for (int i = 0; i < m_pBlocks.size(); ++i)
+	for (int i = 0; i < static_cast<int>(m_pBlocks.size()); ++i)
 	{
 		if (m_pPlayerEnemy && !m_pPlayerEnemy->IsDestroyed())
 		{
